requirements/requirements_image_background_removal.txt
rembg==1.0.7
numpy==1.19.5
Pillow==8.3.1
